+++ common/src/main/java/com/railwayteam/railways/mixin/MixinStationBlock.java
@@ -18,12 +18,12 @@
 
 package com.railwayteam.railways.mixin;
 
-import com.railwayteam.railways.Railways;
 import com.railwayteam.railways.config.CRConfigs;
 import com.railwayteam.railways.content.conductor.ConductorEntity;
 import com.railwayteam.railways.mixin_interfaces.ICarriageConductors;
 import com.railwayteam.railways.registry.CRBlocks;
 import com.railwayteam.railways.registry.CREntities;
+import com.simibubi.create.AllPackets;
 import com.simibubi.create.Create;
 import com.simibubi.create.content.kinetics.deployer.DeployerFakePlayer;
 import com.simibubi.create.content.trains.entity.Carriage;
@@ -35,6 +35,8 @@ import com.simibubi.create.content.trains.schedule.destination.DestinationInstru
 import com.simibubi.create.content.trains.station.GlobalStation;
 import com.simibubi.create.content.trains.station.StationBlock;
 import com.simibubi.create.content.trains.station.StationBlockEntity;
+import com.simibubi.create.content.trains.station.TrainEditPacket;
+import com.simibubi.create.foundation.utility.Components;
 import com.simibubi.create.foundation.utility.VecHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
@@ -49,6 +51,7 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.NameTagItem;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
@@ -259,5 +262,32 @@ public abstract class MixinStationBlock {
         itemEntity.setDeltaMovement(Vec3.ZERO);
         te.getLevel()
             .addFreshEntity(itemEntity);
+    }
+
+    @Inject(method = "use", at = @At("HEAD"), cancellable = true, remap = true)
+    private void deployersNameTag(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer, InteractionHand pHand, BlockHitResult pHit, CallbackInfoReturnable<InteractionResult> cir) {
+        ItemStack itemInHand = pPlayer.getItemInHand(pHand);
+        if (!pLevel.isClientSide && pPlayer instanceof DeployerFakePlayer
+            && pLevel.getBlockEntity(pPos) instanceof StationBlockEntity stationBe
+            && itemInHand.getItem() instanceof NameTagItem
+        ) {
+            cir.setReturnValue(InteractionResult.CONSUME);
+            GlobalStation station = stationBe.getStation();
+            if (station == null || station.getPresentTrain() == null) return;
+
+            Train train = station.getPresentTrain();
+            if (itemInHand.hasCustomHoverName()) { // Set the train name
+                String newName = itemInHand.getHoverName().getString();
+                if (train.name.getString().equals(newName)) return;
+
+                train.name = Components.literal(newName);
+                AllPackets.getChannel().sendToClientsInServer(
+                    new TrainEditPacket.TrainEditReturnPacket(train.id, newName, train.icon.getId()),
+                    pLevel.getServer()
+                );
+            } else { // Get the train's name and put it on the nametag
+                itemInHand.setHoverName(Components.literal(train.name.getString()));
+            }
+        }
     }
 }